some minor fixes & code cleanups (#6)
* fix additional pointer size issues

* fix issue here with overlapping buffer and argument

* add missing #include

* fix some more misc warnings
* buffer/arg overlap
* missing header file <time.h> in main.c
* need to use socklen_t

* current warning settings

* delete ident related code & tarball

it wasn't configured to be on

nobody's really used ident since pervasive NAT, let alone with the
modern approach to privacy.

* this LINKCOUNT macro isn't being defined for some reason

* get rid of more K&R C

* fix issue with my overlapping strings fix - buffer2 was left uninitialised when object wasn't carried

+++ src/mkdata.c
@@ -759,10 +759,10 @@ get_ref (char t, XZON * zo, FILE * F)
   return 0;
 }
 
-int
-get_int (FILE * F)
+intptr_t
+get_int(FILE * F)
 {
-  int i;
+  intptr_t i;
   int c;
   Boolean neg;
 
@@ -1991,7 +1991,7 @@ write_loc (XZON * ZON, int numz, int l)
 	  y = DOOR + ((XOBJ *) L->exits[x])->obj;
 	  break;
 	case '#':
-	  y = (int) (L->exits[x]);
+	  y = (intptr_t) (L->exits[x]);
 	  break;
 	case ' ':
 	  y = -(L->exits[x]->loc);
